# app/Models/Empleado.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

class Empleado extends Model
{
    protected $table = 'empleados';
    protected $primaryKey = 'id_empleado';
    public $timestamps = false;

    protected $fillable = [
        'nombre_completo',
        'rol',
        'activo',
    ];

    protected $casts = [
        'activo' => 'boolean',
    ];

    // Relationships
    public function pedidos()
    {
        return $this->hasMany(Pedido::class, 'id_empleado', 'id_empleado');
    }
}