# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TasteTracker API",
        "description": "API REST del restaurante para gestionar menú, pedidos, clientes y empleados.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/v1/empleados": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Listar todos los empleados",
                "description": "Obtiene el listado completo de empleados ordenado por nombre.",
                "operationId": "389a1db12dba3f5baac5a7e86eaf194a",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Empleado"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Crear un nuevo empleado",
                "description": "Crea un nuevo registro de empleado.",
                "operationId": "30397d9c3a73c4a3d6745d291758328b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpleadoCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empleado creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empleado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/v1/empleados/{empleado}": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Obtener un empleado",
                "description": "Devuelve los detalles de un empleado específico.",
                "operationId": "bf11d45c3ab91aa206b3ad3c06fe4d96",
                "parameters": [
                    {
                        "name": "empleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empleado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualizar un empleado",
                "description": "Actualiza los datos de un empleado.",
                "operationId": "66d5551460926a9fb33944b175866d06",
                "parameters": [
                    {
                        "name": "empleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpleadoUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empleado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Eliminar un empleado",
                "description": "Elimina un registro de empleado.",
                "operationId": "8c91ee03b3848ac67a1fa372d54d5738",
                "parameters": [
                    {
                        "name": "empleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualizar parcialmente un empleado",
                "description": "Actualiza uno o más campos de un empleado.",
                "operationId": "e6ba4d6516fe574f5a9449d32f52f09f",
                "parameters": [
                    {
                        "name": "empleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpleadoUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empleado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empleado": {
                "required": [
                    "nombre_completo",
                    "rol"
                ],
                "properties": {
                    "id_empleado": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre_completo": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "rol": {
                        "type": "string",
                        "example": "Mesero"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "EmpleadoCreate": {
                "required": [
                    "nombre_completo",
                    "rol"
                ],
                "properties": {
                    "nombre_completo": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Juan Pérez"
                    },
                    "rol": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Mesero"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "EmpleadoUpdate": {
                "properties": {
                    "nombre_completo": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Juan P. Pérez"
                    },
                    "rol": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Cajero"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}